refactor: 유저 예약 조회 책임을 예약 서비스로 이동
