+++ src/main/java/roomescape/application/ReservationService.java
@@ -11,6 +11,7 @@ import roomescape.domain.theme.ThemeRepository;
 import roomescape.domain.timeslot.TimeSlot;
 import roomescape.domain.timeslot.TimeSlotRepository;
 import roomescape.domain.user.User;
+import roomescape.domain.user.UserRepository;
 import roomescape.exception.AlreadyExistedException;
 import roomescape.exception.NotFoundException;
 import roomescape.infrastructure.ReservationSpecifications;
@@ -21,15 +22,18 @@ public class ReservationService {
     private final ReservationRepository reservationRepository;
     private final TimeSlotRepository timeSlotRepository;
     private final ThemeRepository themeRepository;
+    private final UserRepository userRepository;
 
     public ReservationService(
             final ReservationRepository reservationRepository,
             final TimeSlotRepository timeSlotRepository,
-            final ThemeRepository themeRepository
+            final ThemeRepository themeRepository,
+            final UserRepository userRepository
     ) {
         this.reservationRepository = reservationRepository;
         this.timeSlotRepository = timeSlotRepository;
         this.themeRepository = themeRepository;
+        this.userRepository = userRepository;
     }
 
     public Reservation reserve(final User user, final LocalDate date, final long timeId, final long themeId) {
@@ -66,5 +70,12 @@ public class ReservationService {
         if (reservation.isPresent()) {
             throw new AlreadyExistedException("이미 예약된 날짜, 시간, 테마에 대한 예약은 불가능합니다.");
         }
+    }
+
+    public List<Reservation> getReservations(final long userId) {
+        var user = userRepository.findById(userId)
+                .orElseThrow(() -> new NotFoundException("존재하지 않는 사용자입니다. id : " + userId));
+
+        return reservationRepository.findByUserId(user.id());
     }
 }
